+++ src/web/components/cards/AssetCard.tsx
@@ -9,6 +9,8 @@ import {
 } from "@mui/material";
 import {
   type Config,
+  floor,
+  getPrecision,
   SHORT_MASK,
   type TradeMemo,
   TradeState,
@@ -38,6 +40,7 @@ const AssetCard = ({ cfg, tm, hideBalances }: Params) => {
   const tradeState = tm.getState();
   const isSold = tradeState === TradeState.SOLD;
 
+  const displayImbalance = tm.imbalanceThreshold() - tm.supplyDemandImbalance;
   return (
     <BasicCard>
       <Box display="flex" justifyContent="space-between" alignItems="center">
@@ -70,6 +73,20 @@ const AssetCard = ({ cfg, tm, hideBalances }: Params) => {
         </Typography>
       </Box>
       <Typography color="text.secondary" variant="body2" mt={1}>
+        {!!tm.currentValue && (
+          <Box
+            display="flex"
+            justifyContent="space-between"
+            alignItems="center"
+          >
+            <Typography variant="inherit" fontWeight="bold" mr={`5px`}>
+              Entry price:
+            </Typography>
+            <Typography variant="inherit">
+              {floor(tm.tradeResult.entryPrice, getPrecision(tm.currentPrice))}
+            </Typography>
+          </Box>
+        )}
         <Box display="flex" justifyContent="space-between" alignItems="center">
           <Typography variant="inherit" fontWeight="bold" mr={`5px`}>
             Paid:
@@ -97,8 +114,8 @@ const AssetCard = ({ cfg, tm, hideBalances }: Params) => {
             </Typography>
             <ImbalanceChecker
               coinName={coinName}
-              initialValue={tm.supplyDemandImbalance}
-              formatter={(v: number) => v + 0.5}
+              initialValue={displayImbalance}
+              formatter={(v: number) => Math.max(0, Math.min(1, v + 0.5))}
             />
           </Box>
         )}